rename rb file
